refactor(ui): extract project feature initializer from app.js
Move project heading functions into client/features/projects/initProjectsFeature.js:
- moveProjectHeading, reorderProjectHeadings, moveTodoToHeading

The feature initializer registers its own window bridge and hooks.
app.js calls initProjectsFeature() during init.

app.js: 1,538 → 1,439 lines (99 lines extracted in this PR)
Total reduction: 1,797 → 1,439 (358 lines, 20%)

Closes #388
