+++ client/app.js
@@ -453,6 +453,7 @@ import {
   registerServiceWorker,
 } from "./bootstrap/initGlobalListeners.js";
 import { initTodosFeature } from "./features/todos/initTodosFeature.js";
+import { initProjectsFeature } from "./features/projects/initProjectsFeature.js";
 import {
   initOnboarding,
   isOnboardingActive,
@@ -662,106 +663,8 @@ const { ensureTodosShellActive, selectWorkspaceView, switchView } =
     prepareFeedbackView,
   });
 
-function moveProjectHeading(headingId, direction) {
-  const headings = getProjectHeadings();
-  const currentIndex = headings.findIndex(
-    (heading) => String(heading.id) === String(headingId),
-  );
-  if (currentIndex < 0) return;
-  const nextIndex = currentIndex + Number(direction);
-  if (nextIndex < 0 || nextIndex >= headings.length) return;
-  const targetId = String(headings[nextIndex]?.id || "");
-  if (!targetId) return;
-  reorderProjectHeadings(String(headingId), targetId, "before");
-}
-function reorderProjectHeadings(draggedId, targetId, placement = "before") {
-  const projectName = getSelectedProjectKey();
-  const projectRecord = getProjectRecordByName(projectName);
-  if (!projectRecord?.id) return;
-
-  const projectId = String(projectRecord.id);
-  const headings = [...getProjectHeadings(projectName)];
-  if (!headings.length) return;
-
-  const draggedIndex = headings.findIndex(
-    (heading) => String(heading.id) === String(draggedId),
-  );
-  const targetIndex = headings.findIndex(
-    (heading) => String(heading.id) === String(targetId),
-  );
-  if (draggedIndex < 0 || targetIndex < 0 || draggedIndex === targetIndex) {
-    return;
-  }
-
-  const [draggedHeading] = headings.splice(draggedIndex, 1);
-  let insertIndex = targetIndex + (placement === "after" ? 1 : 0);
-  if (draggedIndex < insertIndex) {
-    insertIndex -= 1;
-  }
-  insertIndex = Math.max(0, Math.min(insertIndex, headings.length));
-  headings.splice(insertIndex, 0, draggedHeading);
-
-  state.projectHeadingsByProjectId.set(
-    projectId,
-    headings.map((heading, index) => ({
-      ...heading,
-      sortOrder: index,
-    })),
-  );
-  renderTodos();
-
-  void (async () => {
-    try {
-      const response = await apiCall(
-        `${API_URL}/projects/${encodeURIComponent(projectId)}/headings/reorder`,
-        {
-          method: "PUT",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify(
-            headings.map((heading, index) => ({
-              id: String(heading.id),
-              sortOrder: index,
-            })),
-          ),
-        },
-      );
-      if (!response || !response.ok) {
-        throw new Error("Failed to persist heading order");
-      }
-    } catch (error) {
-      console.error("Persist heading reorder failed:", error);
-      await loadHeadingsForProject(projectName);
-      renderTodos();
-      showMessage("todosMessage", "Failed to reorder headings", "error");
-    }
-  })();
-}
-
-async function moveTodoToHeading(todoId, headingIdValue) {
-  const todo = state.todos.find((item) => item.id === todoId);
-  if (!todo) return;
-  const nextHeadingId =
-    typeof headingIdValue === "string" && headingIdValue.trim()
-      ? headingIdValue.trim()
-      : null;
-  try {
-    const updated = await applyTodoPatch(todoId, { headingId: nextHeadingId });
-    if (!nextHeadingId) {
-      renderTodos();
-      return;
-    }
-    const projectName = normalizeProjectPath(
-      updated?.category || todo.category || "",
-    );
-    if (projectName) {
-      await loadHeadingsForProject(projectName);
-    }
-    renderTodos();
-  } catch (error) {
-    console.error("Move todo heading failed:", error);
-    showMessage("todosMessage", "Failed to move task to heading", "error");
-  }
-}
+// moveProjectHeading, reorderProjectHeadings, moveTodoToHeading
+// moved to features/projects/initProjectsFeature.js
 
 // ========== PHASE B: PRIORITY, NOTES, SUBTASKS ==========
 
@@ -1272,8 +1175,7 @@ function bindDockHandlers() {
   hooks.buildHomeTileListByKey = buildHomeTileListByKey;
   // New module hooks (task 149)
   hooks.hideMessage = hideMessage;
-  hooks.moveTodoToHeading = moveTodoToHeading;
-  hooks.reorderProjectHeadings = reorderProjectHeadings;
+  // hooks.moveTodoToHeading, hooks.reorderProjectHeadings — set by initProjectsFeature
   hooks.openTodoDrawer = openTodoDrawer;
   hooks.clearHomeListDrilldown = clearHomeListDrilldown;
   hooks.impactRankForSurface = impactRankForSurface;
@@ -1400,8 +1302,7 @@ window.openEditTodoFromKebab = openEditTodoFromKebab;
 window.openDrawerDangerZone = openDrawerDangerZone;
 window.openTodoFromHomeTile = openTodoFromHomeTile;
 window.moveTodoToProject = moveTodoToProject;
-window.moveTodoToHeading = moveTodoToHeading;
-window.moveProjectHeading = moveProjectHeading;
+// moveTodoToHeading, moveProjectHeading — registered by initProjectsFeature
 // Drag and drop
 window.handleDragStart = DragDrop.handleDragStart;
 window.handleDragOver = DragDrop.handleDragOver;
@@ -1471,6 +1372,7 @@ window.runAdminFeedbackAutomation = runAdminFeedbackAutomation;
 // ---------------------------------------------------------------------------
 function init() {
   initTodosFeature();
+  initProjectsFeature();
   bindResponsiveLayoutState();
   renderSidebarNavigation();
   bindCriticalHandlers();
